a12_system/requirements.txt
opencv-python-headless>=4.8,<5
requests>=2.31,<3
numpy>=1.24,<2
paho-mqtt>=2.0,<3
python-dotenv>=1.0,<2
Pillow>=10.0,<11
pyTelegramBotAPI>=4.14,<5
imageio>=2.31,<3
face_recognition>=1.3,<2
